feat: app mobile ClassLink (Flutter) + super-admin amélioré
Super-admin :
- Suppression d'établissement avec confirmation par nom
- Remises promotionnelles (%) sur les plans tarifaires, appliquées à l'inscription

API REST mobile (Next.js) :
- Auth JWT : login, refresh
- Notes, emploi du temps, absences, annonces, cantine, messages
- Parent : liste enfants, détail enfant
- Paiements scolaires (élève & parent)
- Bulletins PDF par trimestre
- FCM : enregistrement token appareil

App mobile Flutter (classlink_mobile) :
- Login école + email + mot de passe
- Dashboard élève (moyenne générale) et parent
- Notes par matière avec moyennes
- Emploi du temps par jour (lundi–vendredi)
- Absences avec statistiques
- Annonces, cantine, messagerie
- Bulletins : liste trimestres + viewer PDF (génération Dart, partage natif)
- Paiements avec statuts colorés
- Espace parent : suivi complet de chaque enfant
- Notifications push FCM (infrastructure prête, activable via Firebase)
- Nom de l'app : ClassLink

+++ classelink/actions/teacher.ts
@@ -90,7 +90,7 @@ export async function getStudentsWithGrades(
         '[]'::json
       ) AS grades,
       CASE WHEN COUNT(g.id) > 0
-        THEN ROUND(SUM(g.value * g.coefficient) / NULLIF(SUM(g.coefficient), 0), 2)
+        THEN (ROUND(SUM(g.value * g.coefficient) / NULLIF(SUM(g.coefficient), 0), 2))::float8
         ELSE NULL
       END AS average
     FROM enrollments e
